requirements.txt
streamlit
python-dotenv
langchain
faiss-cpu
youtube-transcript-api
google-generativeai